requirements.txt
Flask==3.1.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.5
keyring==25.6.0
keyrings.alt==5.0.2
